make compile

+++ ros/src/twist_controller/src/MPC.cpp
@@ -51,17 +51,17 @@ public:
         // Reference State Cost
         // TODO: Define the cost related the reference state and
         // any anything you think may be beneficial.
-        for (int i = 0; i < N; i++) {
+        for (size_t i = 0; i < N; i++) {
             fg[0] += 3000 * CppAD::pow(vars[cte_start + i], 2);
             fg[0] += 3000 * CppAD::pow(vars[epsi_start + i], 2);
             fg[0] += 1 * CppAD::pow(vars[v_start + i] - ref_v, 2);
         }
 
-        for (int i = 0; i < N - 1; i++) {
+        for (size_t i = 0; i < N - 1; i++) {
             fg[0] += 1000 * CppAD::pow(vars[delta_start + i], 2);
             fg[0] += 5 * CppAD::pow(vars[a_start + i], 2);
         }
-        for (int t = 0; t < N - 2; t++) {
+        for (size_t t = 0; t < N - 2; t++) {
             fg[0] += 5000 * CppAD::pow(vars[delta_start + t + 1] - vars[delta_start + t], 2);
             fg[0] += 10 * CppAD::pow(vars[a_start + t + 1] - vars[a_start + t], 2);
         }
@@ -84,7 +84,7 @@ public:
         fg[1 + epsi_start] = vars[epsi_start];
 
         // The rest of the constraints
-        for (int t = 1; t < N; t++) {
+        for (size_t t = 1; t < N; t++) {
             AD<double> x1 = vars[x_start + t];
             AD<double> x0 = vars[x_start + t - 1];
             AD<double> y1 = vars[y_start + t];
@@ -150,7 +150,7 @@ vector<double> MPC::Solve(Eigen::VectorXd state, Eigen::VectorXd coeffs) {
     // Initial value of the independent variables.
     // SHOULD BE 0 besides initial state.
     Dvector vars(n_vars);
-    for (int i = 0; i < n_vars; i++) {
+    for (size_t i = 0; i < n_vars; i++) {
         vars[i] = 0;
     }
 
@@ -165,18 +165,18 @@ vector<double> MPC::Solve(Eigen::VectorXd state, Eigen::VectorXd coeffs) {
     vars[cte_start] = cte;
     vars[epsi_start] = epsi;
 
-    for (int i = 0; i < delta_start; i++) {
+    for (size_t i = 0; i < delta_start; i++) {
         vars_lowerbound[i] = -1.0e19;
         vars_upperbound[i] = 1.0e19;
     }
-    for (int i = delta_start; i < a_start; i++) {
+    for (size_t i = delta_start; i < a_start; i++) {
         vars_lowerbound[i] = -0.436332 ;
         vars_upperbound[i] = 0.436332;
     }
 
     // Acceleration/decceleration upper and lower limits.
     // NOTE: Feel free to change this to something else.
-    for (int i = a_start; i < n_vars; i++) {
+    for (size_t i = a_start; i < n_vars; i++) {
         vars_lowerbound[i] = -1.0;
         vars_upperbound[i] = 1.0;
     }
@@ -185,7 +185,7 @@ vector<double> MPC::Solve(Eigen::VectorXd state, Eigen::VectorXd coeffs) {
     // Should be 0 besides initial state.
     Dvector constraints_lowerbound(n_constraints);
     Dvector constraints_upperbound(n_constraints);
-    for (int i = 0; i < n_constraints; i++) {
+    for (size_t i = 0; i < n_constraints; i++) {
         constraints_lowerbound[i] = 0;
         constraints_upperbound[i] = 0;
     }
@@ -248,7 +248,7 @@ vector<double> MPC::Solve(Eigen::VectorXd state, Eigen::VectorXd coeffs) {
     result.push_back(solution.x[delta_start]);
     result.push_back(solution.x[a_start]);
 
-    for (int i = 0; i < N - 1; i++) {
+    for (size_t i = 0; i < N - 1; i++) {
         result.push_back(solution.x[x_start + i + 1]);
         result.push_back(solution.x[y_start + i + 1]);
     }
